refactor all getters for gated nfts in the same hook
